Refactor layout viewer panel modules
